requirements.txt
pandas==2.2.1
openai==1.14.0
streamlit==1.32.0
plotly==5.20.0
scipy==1.12.0
pyarrow==15.0.1
python-dotenv==1.0.1
